Transformations improved

+++ src/components/canvas/CanvasMapper.java
@@ -1,19 +1,19 @@
 package components.canvas;
 
 public class CanvasMapper {
-    public static int mapXjToXp(int xj, int width) {
+    public static double mapXjToXp(double xj, double width) {
         return xj - (width / 2);
     }
 
-    public static int mapYjToYp(int yj, int height) {
+    public static double mapYjToYp(double yj, double height) {
         return (height / 2) - yj;
     }
 
-    public static int mapYpToYj(int yp, int height) {
+    public static double mapYpToYj(double yp, double height) {
         return (height / 2) - yp;
     }
 
-    public static int mapXpToXj(int xp, int width) {
+    public static double mapXpToXj(double xp, double width) {
         return xp + (width / 2);
     }
 }
